Added tests for problem class.
The problem class now can take in both the seperated and combined constraints.
The tests seem to pass. Need to add functionality (and tests!) for simple
bounds on x for the seperated case. Also should check for combined case:
if only equality or on inequality constraints are supplied, that the other
output function (h or g in this case) is None.
Also, need to reconsider changing functions g, h to c_eq, c_ineq.