Minor refinements to QP, added initial SQP
The S-C approach made the QP solver slower, so now it just is doing a direct
solve. Following this commit, the dense option is going to go away - it's not
worth the headache of maintaining two pathways.

The SQP solver has been coded but not thoroughly tested. "Stalling" in the QP
solver is preventing this. Next steps are fix stalling w/ perturbation of
bounds, remove dense option, and (possibly) investigate null space method for
QP solver.